use filter and moved function to different file to avoid pegjs linting issues
